backend/chat-service/requirements.txt
# ============================================================
# Chat Service - Dependencies
# ============================================================

# FastAPI & Server
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9

# Pydantic
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# JWT (for validating auth-service tokens)
python-jose[cryptography]>=3.3.0

# LangChain & LangGraph
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0
langgraph>=0.2.0
langgraph-checkpoint>=2.0.0
langchain-mcp-adapters>=0.1.0

# E2B Code Interpreter
e2b-code-interpreter>=1.0.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
